[Refactor] CustomSlider 리팩토링 및 Filter 내부 폴더 구조 변경
